Moved type inference tests to separate file
